pyproject.toml
[tool.black]
line-length = 120
target-version = ['py38']

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 120
multi_line_output = 3
profile = "black"
py_version = 38
use_parentheses = true

[tool.poetry]
name = "nfcli"
version = "1.0.0"
description = "A CLI tool to pretty print fleet and ship files to console."
authors = ["Jaghaimo <1764586+jaghaimo@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
xmltodict = "^0.13.0"
rich = "^12.4.4"
"discord.py" = "^1.7.3"
python-dotenv = "^0.20.0"
CairoSVG = "^2.5.2"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.12.2"
steam = "^1.3.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[tool.poetry.scripts]
nfcli = "nfcli.__main__:main"
"bot.upload" = "nfcli.bot.upload:start"
"bot.wiki" = "nfcli.bot.wiki:start"
"bot.workshop" = "nfcli.bot.workshop:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
